Fix functions with dots in names

+++ data/typing/numpy.lib.stride_tricks.py
@@ -1,4 +1,4 @@
 
 
-def as_strided(x: numpy.ndarray, shape: tuple[(Union[(numpy.int64, int)], ...)], strides: tuple[(Union[(int, numpy.int64)], ...)]):
+def as_strided(x: numpy.ndarray, shape: tuple[(Union[(int, numpy.int64)], ...)], strides: tuple[(Union[(int, numpy.int64)], ...)]):
     'usage.skimage: 11'
